Add tag minimization utility and hook auto AI pipeline
